+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `localisationService.getIOSVersion()`
+
 ### Changed
 
 - CP Pay redirect to handle v2 query parameters